+++ app/_components/landing/BentoSection/components/index.ts
@@ -1,0 +1,2 @@
+export * from './GradientBlur';
+export * from './OrbitingStack';
