+++ app/_components/landing/FAQSection/constants/index.ts
@@ -1,0 +1,14 @@
+export const FAQ_ITEMS = [
+  {
+    question: 'page.home.section.faq.0.question',
+    answer: 'page.home.section.faq.0.answer'
+  },
+  {
+    question: 'page.home.section.faq.1.question',
+    answer: 'page.home.section.faq.1.answer'
+  },
+  {
+    question: 'page.home.section.faq.2.question',
+    answer: 'page.home.section.faq.2.answer'
+  }
+] as const;
