+++ app/_components/landing/ReviewsSection/constants/index.ts
@@ -1,0 +1,28 @@
+export const REVIEWS = [
+  {
+    name: 'John Doe',
+    role: 'Software Engineer',
+    review:
+      'This bootcamp was an amazing experience! The curriculum was well-structured and the instructors were very knowledgeable. '
+  },
+  {
+    name: 'Jane Smith',
+    role: 'Frontend Developer',
+    review: 'I highly recommend this bootcamp to anyone looking to break into the tech industry. '
+  },
+  {
+    name: 'Emily Johnson',
+    role: 'Full Stack Developer',
+    review: 'The support from the community and mentors was incredible. '
+  },
+  {
+    name: 'Michael Brown',
+    role: 'Data Scientist',
+    review: 'The bootcamp provided a great balance of theory and practical application. '
+  },
+  {
+    name: 'Sarah Wilson',
+    role: 'UX Designer',
+    review: 'I appreciated the focus on both technical skills and soft skills. '
+  }
+] as const;
