+++ app/_components/landing/index.ts
@@ -1,0 +1,5 @@
+export * from './BentoSection/BentoSection';
+export * from './FAQSection/FAQSection';
+export * from './HeroSection/HeroSection';
+export * from './ReviewsSection/ReviewsSection';
+export * from './SubfooterSection/SubfooterSection';
