+++ app/_components/layout/CookieConsent/scripts/index.ts
