+++ app/_components/layout/index.ts
@@ -1,5 +1,3 @@
-export * from './Bento/GradientBlur';
-export * from './Bento/OrbitingStack';
 export * from './CookieConsent/DynamicCookieConsent';
 export * from './Footer/Footer';
 export * from './Header/Header';
