+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { Nunito, Overpass_Mono, Parisienne, Pixelify_Sans } from 'next/font/goog
 import { intl } from '@/intl/server';
 import { cn } from '@/lib/utils';
 
-import { DynamicCookieConsent, Footer, Header } from './_components';
+import { DynamicCookieConsent, Footer, Header } from './_components/layout';
 import { ThemeScript } from './_scripts';
 import { Provider } from './provider';
