+++ src/components/common/motion/index.ts
@@ -1,0 +1,2 @@
+export * from './motion-card';
+export * from './motion-matrix-grid';
