+++ src/components/common/motion/motion-matrix-grid.tsx
@@ -2,6 +2,6 @@
 
 import { motion } from 'motion/react';
 
-import { MatrixGrid } from './matrix-grid';
+import { MatrixGrid } from '../matrix-grid';
 
 export const MotionMatrixGrid = motion.create(MatrixGrid);
